Refactoring: separate functions by module role
