+++ public/index.html
@@ -77,7 +77,5 @@
         </div>
       </div>
     </div>
-    
   </body>
-  <script src="../dist/bundle.js"></script>
 </html>
