+++ public/js/component/themecomponent.js
@@ -1,10 +1,45 @@
-import {myDomApi} from "./mydomapi.js"
-import {themeData} from "./index.js"
+import {URL} from "../url.js"
+import {myDomApi} from "../util/mydomapi.js"
 
+let themeData;
+let themeLength = 5;
+const themeImgCnt = 5;
+const extendCnt = 5;
 const moreBtn = myDomApi.myQuerySelector("#more");
 
-let themeLength = 5;
-const extendCnt = 5;
+const createThemeContainer = () => {
+  let themeContainer = myDomApi.myQuerySelector("table.theme-container");
+  let newLayout = "<tbody style='display:inline-block'>";
+  for(let idx=0; idx<themeImgCnt; idx++) {
+    newLayout += `
+      <th class="theme">
+        <img class="theme-img">
+        <div class="theme-title"></div>
+        <div class="theme-info"></div>
+        <img class="theme-icon" src="img/themeIcon.png"></img>
+      </th>
+    `
+  }
+  themeContainer.innerHTML += newLayout + "</tbody>";
+
+  let themeImg = myDomApi.myQuerySelectorAll("img.theme-img");
+  let themeTitle = myDomApi.myQuerySelectorAll("div.theme-title");
+  let themeInfo = myDomApi.myQuerySelectorAll("div.theme-info");
+
+  const request = new Request(URL + "/theme");
+  fetch(request)
+  .then(response => response.text())
+  .then(result => {
+    themeData = JSON.parse(result)["items"];
+    for(let idx=0; idx<themeImgCnt; idx++){
+      themeImg[idx].src = themeData[idx].src;
+      themeTitle[idx].innerHTML = themeData[idx].title;
+      themeInfo[idx].innerHTML = themeData[idx].subtitle;
+    }
+  })
+  .catch(error => console.log('error', error));
+  themeClickListener();
+}
 
 const extendLayout = () => {
   let themeContainer = myDomApi.myQuerySelector("table.theme-container");
@@ -50,4 +85,4 @@ const themeClickListener = () => {
 
 moreBtn.addEventListener("click", extendLayout);
 
-export {themeClickListener};
+export {createThemeContainer}
