+++ public/js/url.js
@@ -1,0 +1,3 @@
+const URL = "https://7aebe337-b81c-42de-b89f-8c268823df03.mock.pstmn.io";
+
+export {URL}
